requirements.txt
pandas==2.0.3
requests
sqlalchemy==1.4.52
psycopg2-binary
python-dotenv